# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Coder API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  userId: number;
  messageCount: number;
  fileCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  projectId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface GeneratedFile {
  id: number;
  projectId: number;
  messageId: number;
  filename: string;
  fileType: string;
  size: number;
  createdAt: string;
}

export interface MessageResponse {
  userMessage: Message;
  assistantMessage: Message;
  generatedFiles: GeneratedFile[];
}

export interface UserStats {
  projectCount: number;
  messageCount: number;
  fileCount: number;
}