# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Coder API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: projects
    description: Project operations
  - name: messages
    description: Chat message operations
  - name: files
    description: File generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: OK

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user projects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  /projects/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /projects/{id}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List generated files in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedFile"

  /files/{fileId}/download:
    get:
      operationId: downloadFile
      tags: [files]
      summary: Download a generated file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  /projects/stats:
    get:
      operationId: getStats
      tags: [projects]
      summary: Get user stats (project count, message count, file count)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - createdAt

    RegisterInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 6
      required:
        - username
        - email
        - password

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        userId:
          type: integer
        messageCount:
          type: integer
        fileCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - userId
        - messageCount
        - fileCount
        - createdAt
        - updatedAt

    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - name

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    Message:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - role
        - content
        - createdAt

    MessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content

    MessageResponse:
      type: object
      properties:
        userMessage:
          $ref: "#/components/schemas/Message"
        assistantMessage:
          $ref: "#/components/schemas/Message"
        generatedFiles:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedFile"
      required:
        - userMessage
        - assistantMessage
        - generatedFiles

    GeneratedFile:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        messageId:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        size:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - projectId
        - messageId
        - filename
        - fileType
        - size
        - createdAt

    UserStats:
      type: object
      properties:
        projectCount:
          type: integer
        messageCount:
          type: integer
        fileCount:
          type: integer
      required:
        - projectCount
        - messageCount
        - fileCount