# lib/api-zod/src/generated/types/generatedFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Coder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedFile {
  id: number;
  projectId: number;
  messageId: number;
  filename: string;
  fileType: string;
  size: number;
  createdAt: string;
}