# lib/api-zod/src/generated/types/messageResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Coder API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedFile } from './generatedFile';
import type { Message } from './message';

export interface MessageResponse {
  userMessage: Message;
  assistantMessage: Message;
  generatedFiles: GeneratedFile[];
}