# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Coder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  userId: number;
  messageCount: number;
  fileCount: number;
  createdAt: string;
  updatedAt: string;
}